requirements.txt
Flask
Flask-CORS
psycopg2-binary
pandas
matplotlib
seaborn
plotly